refactor: Move inference models to backend subdirectory
Move all AI inference models and setup files from root-level
inference_models/ directory to backend/inference_models/ to
consolidate the backend codebase into a single directory.

- Move ASR, slide matching, and TTS processors
- Move pipeline orchestration code
- Move environment configuration and setup script
