+++ backend/environment.yml
@@ -1,0 +1,83 @@
+name: swpp-backend
+channels:
+  - conda-forge
+dependencies:
+  - python=3.12
+  - numpy=2.3.3
+  - opencv
+  - pillow
+  - ninja
+  - pip
+  - pip:
+    # PyTorch with CUDA support
+    - --index-url https://download.pytorch.org/whl/cu129
+    - --extra-index-url https://pypi.org/simple
+    - torch==2.8.0
+    - torchvision==0.23.0
+    - torchaudio==2.8.0
+    - cuda-python==12.9
+
+    # ML/AI libraries
+    - transformers==4.51.3
+    - tokenizers
+    - datasets
+    - peft
+    - accelerate
+    - bitsandbytes
+    - safetensors
+    - kokoro==0.9.4
+    - soundfile
+    - librosa
+    - 'nemo_toolkit[asr]==2.4.1'
+    - openpyxl
+    - pymupdf
+    - tqdm
+
+    # Backend/API libraries
+    - aiofiles==25.1.0
+    - annotated-types==0.7.0
+    - anyio==4.11.0
+    - certifi==2025.10.5
+    - click==8.3.0
+    - dnspython==2.8.0
+    - email-validator==2.3.0
+    - fastapi==0.118.3
+    - fastapi-cli==0.0.13
+    - fastapi-cloud-cli==0.3.1
+    - h11==0.16.0
+    - httpcore==1.0.9
+    - httptools==0.7.1
+    - httpx==0.28.1
+    - idna==3.10
+    - itsdangerous==2.2.0
+    - Jinja2==3.1.6
+    - markdown-it-py==4.0.0
+    - MarkupSafe==3.0.3
+    - mdurl==0.1.2
+    - pydantic==2.12.0
+    - pydantic_core==2.41.1
+    - Pygments==2.19.2
+    - python-dotenv==1.1.1
+    - python-multipart==0.0.20
+    - PyYAML==6.0.3
+    - requests==2.32.5
+    - rich==14.2.0
+    - rich-toolkit==0.15.1
+    - rignore==0.7.0
+    - sentry-sdk==2.41.0
+    - setuptools==80.9.0
+    - shellingham==1.5.4
+    - sniffio==1.3.1
+    - sseclient-py==1.8.0
+    - sse-starlette==3.0.2
+    - starlette==0.48.0
+    - typer==0.19.2
+    - typing-inspection==0.4.2
+    - typing_extensions==4.15.0
+    - ujson==5.11.0
+    - urllib3==2.5.0
+    - uvicorn==0.37.0
+    - uvloop==0.21.0
+    - watchfiles==1.1.0
+    - websockets==15.0.1
+    - wheel==0.45.1
