--- inference_models/environment.yml
@@ -1,31 +1,0 @@
-name: swpp-ai
-channels:
-  - conda-forge
-dependencies:
-  - python=3.12
-  - numpy=2.3.3
-  - opencv
-  - pillow
-  - ninja
-  - pip
-  - pip:
-    - --index-url https://download.pytorch.org/whl/cu129
-    - --extra-index-url https://pypi.org/simple
-    - torch==2.8.0
-    - torchvision==0.23.0
-    - torchaudio==2.8.0
-    - cuda-python==12.9
-    - transformers==4.51.3
-    - tokenizers
-    - datasets
-    - peft
-    - accelerate
-    - bitsandbytes
-    - safetensors
-    - kokoro==0.9.4
-    - soundfile
-    - librosa
-    - 'nemo_toolkit[asr]==2.4.1'
-    - openpyxl
-    - pymupdf
-    - tqdm
